+++ webknossos/examples/dataset_usage.py
@@ -25,26 +25,22 @@ def main() -> None:
         layer_name="color", category="color", dtype_per_channel="uint8", num_channels=3
     )
     mag1 = layer.add_mag("1")
-    mag1 = layer.add_mag("2")
+    mag2 = layer.add_mag("2")
 
     ##########################
     # Writing into a dataset #
     ##########################
 
-    layer = dataset.get_layer("color")
-    mag1 = layer.get_mag("1")
-    mag2 = layer.get_mag("2")
-
     # The properties are updated automatically
     # when the written data exceeds the bounding box in the properties
     mag1.write(
-        offset=(10, 20, 30),
+        absolute_offset=(10, 20, 30),
         # assuming the layer has 3 channels:
         data=(np.random.rand(3, 512, 512, 32) * 255).astype(np.uint8),
     )
 
     mag2.write(
-        offset=(5, 10, 15),
+        absolute_offset=(10, 20, 30),
         data=(np.random.rand(3, 256, 256, 16) * 255).astype(np.uint8),
     )
 
@@ -52,15 +48,12 @@ def main() -> None:
     # Reading from a dataset #
     ##########################
 
-    layer = dataset.get_layer("color")
-    mag1 = layer.get_mag("1")
-    mag2 = layer.get_mag("2")
-
     data_in_mag1 = mag1.read()  # the offset and size from the properties are used
-    data_in_mag1_subset = mag1.read(offset=(10, 20, 30), size=(512, 512, 32))
+    data_in_mag1_subset = mag1.read(absolute_offset=(10, 20, 30), size=(512, 512, 32))
 
     data_in_mag2 = mag2.read()
-    data_in_mag2_subset = mag2.read(offset=(5, 10, 15), size=(256, 256, 16))
+    data_in_mag2_subset = mag2.read(absolute_offset=(10, 20, 30), size=(512, 512, 32))
+    assert data_in_mag2_subset.shape == (3, 256, 256, 16)
 
     #####################
     # Copying a dataset #